Re-write card shuffling/dealing
The previous way effectively did one shuffle on the deck. There
is a long comment inline in the Shuffle method explaining exactly
why I think that is not sufficient.

As it is not trivial to deterministically and fairly do draws
on the same deck while also shuffling it twice in the same method
and with different seeds and so on, the methods were refactored
to split the concern of extracting a number of cards from a deck
from the concern of doing so randomly.